requirements.txt
sqlmodel
psycopg2-binary
python-dotenv
fastapi
uvicorn
python-jose
python-multipart
requests
